@@ -1,34 +1,24 @@
 {
-  "name": "phoenixhell",
+  "name": "react-router5",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/icons": "latest",
-    "@craco/craco": "^6.2.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "antd": "^4.16.6",
-    "axios": "^0.21.1",
-    "babel-plugin-import": "^1.13.3",
-    "craco-less": "^1.18.0",
+    "@reduxjs/toolkit": "^2.2.1",
+    "antd": "^5.14.1",
+    "axios": "^1.6.7",
     "nanoid": "^3.1.23",
-    "pubsub-js": "^1.9.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.0",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
-    "start": "^5.1.0",
-    "web-vitals": "^1.1.2"
+    "react-scripts": "5.0.1",
+    "redux": "^5.0.1",
+    "redux-thunk": "^3.1.0"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco  build",
-    "test": "craco test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -48,5 +38,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
   }
 }
