@@ -1,6 +1,6 @@
 import {combineReducers} from "redux";
-import countReducer from "./count"
-import personReducer from "./person"
+import count from "./count"
+import person from "./person"
 
 //汇总所有reducer变为一个总的reducers
 /* 
@@ -8,9 +8,16 @@ import personReducer from "./person"
     const mapStateToProps = state => ({count: state.count,person:state.person});
 
 */
+
+// const reducers =combineReducers({
+//     count:countReducer,
+//     person:personReducer
+// });
+
+//将名字改成和key一样,触发简写
 const reducers =combineReducers({
-    count:countReducer,
-    person:personReducer
+    count,
+    person
 });
 
 export default reducers
